servers#destroy fix handling incorrect server ids. Fixes #3565

+++ lib/bbcloud/commands/servers-destroy.rb
@@ -7,10 +7,10 @@ command [:destroy] do |c|
       raise "you must specify the id of the server(s) you want to destroy"
     end
 
-
     servers = args.collect do |sid|
       info "Destroying server #{sid}"
       server = Server.find sid
+      raise "Server #{sid} does not exist" if server.nil?
       begin
         server.destroy
       rescue Brightbox::Api::Conflict => e
